fix(ci): maintainability gate — split codex readiness tests into part_005 to keep part_001 under the 1500-line limit
